refactor: reorganize configuration files and implement global exception handling
