--- config/envs.ts
@@ -1,35 +1,0 @@
-import 'dotenv/config';
-
-import * as joi from 'joi';
-
-interface EnvVars {
-  DB_USER: string;
-  DB_PASSWORD: string;
-  DB_CONNECTION_STRING: string;
-  LD_LIBRARY_PATH: string;
-}
-
-const envsSchema = joi
-  .object({
-    DB_USER: joi.string().required(),
-    DB_PASSWORD: joi.string().required(),
-    DB_CONNECTION_STRING: joi.string().required(),
-    LD_LIBRARY_PATH: joi.string().default('C:/ORACLE/instantclient_12_1'),
-  })
-  .unknown()
-  .required();
-
-const validationSchema = envsSchema.validate(process.env);
-
-if (validationSchema.error) {
-  throw new Error(`Config validation error: ${validationSchema.error.message}`);
-}
-
-const envVars: EnvVars = validationSchema.value as EnvVars;
-
-export const envs = {
-  DB_USER: envVars.DB_USER,
-  DB_PASSWORD: envVars.DB_PASSWORD,
-  DB_CONNECTION_STRING: envVars.DB_CONNECTION_STRING,
-  LD_LIBRARY_PATH: envVars.LD_LIBRARY_PATH,
-};
