+++ src/filters/all-exception.filter.ts
@@ -1,0 +1,9 @@
+import { ArgumentsHost, Catch } from '@nestjs/common';
+import { BaseRpcExceptionFilter } from '@nestjs/microservices';
+
+@Catch()
+export class AllExceptionsFilter extends BaseRpcExceptionFilter {
+  catch(exception: unknown, host: ArgumentsHost) {
+    return super.catch(exception, host);
+  }
+}
